+++ app/src/main/java/ru/lizzzi/crossfit_rekord/DocumentFields_Table.java
@@ -1,47 +1,17 @@
 package ru.lizzzi.crossfit_rekord;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-
-import ru.profit_group.scorocode_sdk.scorocode_objects.DocumentInfo;
-
-//import static prof_itgroup.ru.storehouseapp.Helpers.Helper.getColorsListFrom;
-
-/**
- * Created by Peter Staranchuk on 10/30/16
- */
 
 public class DocumentFields_Table {
     private Context context;
-    private DocumentInfo documentInfo;
 
     public DocumentFields_Table(Context context) {
-        this(context, null);
-    }
-
-    public DocumentFields_Table(Context context, DocumentInfo documentInfo) {
         this.context = context;
-        this.documentInfo = documentInfo;
-    }
-
-    public void setDocumentInfo(DocumentInfo documentInfo) {
-        this.documentInfo = documentInfo;
     }
 
     public enum Fields {
-
-        day_of_week(R.string.day_of_week),
         start_time(R.string.start_time),
         type(R.string.type);
-
-        /*
-        PLATFORM(R.string.fieldPlatformName),
-        CAMERA_INFO(R.string.fieldCameraInfo),
-        COLORS_AVAILABLE(R.string.fieldColorsAvailable),
-        DEVICE_PRICE(R.string.fieldDevicePrice),
-        LAST_SEND(R.string.fieldLastSend),
-        BUYERS(R.string.fieldBuyers),
-        SEND_INFO(R.string.fieldSendInfo);*/
 
         private int fieldNameId;
 
@@ -54,10 +24,6 @@ public class DocumentFields_Table {
         }
 
     }
-    public String getDayOfWeekFields() {
-        return context.getString(Fields.day_of_week.getNameId());
-    }
-
 
     public String getStartTimeField() {
         return context.getString(Fields.start_time.getNameId());
@@ -66,106 +32,4 @@ public class DocumentFields_Table {
     public String getTypeField() {
         return context.getString(Fields.type.getNameId());
     }
-/*
-    public String getCameraInfoField() {
-        return context.getString(Fields.CAMERA_INFO.getNameId());
-    }
-
-    public String getColorsAvailableField() {
-        return context.getString(Fields.COLORS_AVAILABLE.getNameId());
-    }
-
-    public String getDevicePriceField() {
-        return context.getString(Fields.DEVICE_PRICE.getNameId());
-    }
-
-    public String getLastSendField() {
-        return context.getString(Fields.LAST_SEND.getNameId());
-    }
-
-    public String getBuyersField() {
-        return context.getString(Fields.BUYERS.getNameId());
-    }
-
-    public String getSendInfoField() {
-        return context.getString(Fields.SEND_INFO.getNameId());
-    }*/
-
-    private String getFieldValue(DocumentInfo documentInfo, Fields field) {
-        return String.valueOf(documentInfo.getFields().get(context.getString(field.getNameId())));
-    }
-
-    @NonNull
-    public String getTypesOfTraining() {
-        if(documentInfo != null) {
-            return getFieldValue(documentInfo, Fields.day_of_week);
-        }
-        return "";
-    }
-
-    @NonNull
-    public String getPeriod() {
-        if(documentInfo != null) {
-            return getFieldValue(documentInfo, Fields.start_time);
-        }
-        return "";
-    }
-
-    @NonNull
-    public String getPrice() {
-        if(documentInfo != null) {
-            return getFieldValue(documentInfo, Fields.type);
-        }
-        return "";
-    }
-/*
-    @NonNull
-    public String getColors() {
-        if(documentInfo != null) {
-            return getFieldValue(documentInfo, Fields.COLORS_AVAILABLE)
-                    .replace("[", "")
-                    .replace("]", "");
-        }
-        return "";
-    }
-
-    @NonNull
-    public Double getDevicePrice() {
-        if(documentInfo != null) {
-            return Double.valueOf(getFieldValue(documentInfo, Fields.DEVICE_PRICE));
-        }
-        return new Double(0);
-    }
-
-    @NonNull
-    public String getLastSendTime() {
-        String date = getFieldValue(documentInfo, Fields.LAST_SEND);
-        if (date != null && !date.isEmpty() && !date.equals("null")) {
-            SimpleDateFormat formatter = new SimpleDateFormat("dd-MMM-yyyy hh:mm a");
-            return formatter.format(Date.parse(date));
-        }
-
-        return "";
-    }
-
-    @NonNull
-    public String getBuyers() {
-        if (documentInfo != null) {
-            String buyers = getFieldValue(documentInfo, Fields.BUYERS);
-            if(buyers != null) {
-                buyers = buyers.replace("[", "").replace("]", "");
-                if(buyers.isEmpty() || buyers.equals("null")) {
-                    return "";
-                } else {
-                    return buyers;
-                }
-            }
-        }
-        return "";
-    }
-
-    @NonNull
-    public List<String> getColorsAsList() {
-        return getColorsListFrom(getColors());
-    }*/
 }
