+++ app/src/main/java/ru/lizzzi/crossfit_rekord/ListenerActivity.java
@@ -4,6 +4,6 @@ package ru.lizzzi.crossfit_rekord;
  * Created by Liza on 22.11.2017.
  */
 
-interface ListenerActivity {
+public interface ListenerActivity {
     void Remove(String exercises, int position);
 }
