+++ app/src/main/java/ru/lizzzi/crossfit_rekord/adapters/RecyclerAdapterRecord.java
@@ -1,6 +1,7 @@
-package ru.lizzzi.crossfit_rekord;
+package ru.lizzzi.crossfit_rekord.adapters;
 
 import android.content.Context;
+import android.content.SharedPreferences;
 import android.support.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -10,6 +11,8 @@ import android.widget.TextView;
 
 import java.util.List;
 
+import ru.lizzzi.crossfit_rekord.DocumentFields_Record;
+import ru.lizzzi.crossfit_rekord.R;
 import ru.profit_group.scorocode_sdk.scorocode_objects.DocumentInfo;
 
 /*public class RecyclerAdapterMenu extends RecyclerView.Adapter<RecyclerAdapterMenu.ViewHolder> {
@@ -77,19 +80,24 @@ import ru.profit_group.scorocode_sdk.scorocode_objects.DocumentInfo;
     }
     */
 
-public class RecyclerAdapterMenu extends BaseAdapter {
+public class RecyclerAdapterRecord extends BaseAdapter {
     private Context context;
     private List<DocumentInfo> storedItems;
     private int layoutId;
     private LayoutInflater inflater;
-    private DocumentFields fields;
+    private DocumentFields_Record fields;
 
-    public RecyclerAdapterMenu(Context context, @NonNull List<DocumentInfo> storedItems, int layoutId) {
+    public static final String APP_PREFERENCES = "audata";
+    public static final String APP_PREFERENCES_USERCOUNT = "Usercount";
+    SharedPreferences mSettings;
+
+    public RecyclerAdapterRecord(Context context, @NonNull List<DocumentInfo> storedItems, int layoutId) {
         this.context = context;
         this.storedItems = storedItems;
         this.layoutId = layoutId;
         inflater = LayoutInflater.from(context);
-        fields = new DocumentFields(context, null);
+        fields = new DocumentFields_Record(context, null);
+        mSettings = context.getSharedPreferences(APP_PREFERENCES, context.MODE_PRIVATE);
     }
 
     @Override
@@ -126,13 +134,19 @@ public class RecyclerAdapterMenu extends BaseAdapter {
 
     private void customizeView(View view, ViewHolder holder, final DocumentInfo documentInfo) {
 
-        String period = (String) documentInfo.getFields().get(fields.getPeriodField());
-        String typesoftraining = (String) documentInfo.getFields().get(fields.getTypesOfTrainingFields());
-        Double price = (Double) documentInfo.getFields().get(fields.getPriceField());
+        String number = mSettings.getString(APP_PREFERENCES_USERCOUNT, "");
 
-        holder.PeriodItem.setText(period);
-        holder.TypesOfTrainingItem.setText(typesoftraining);
-        holder.PriceItem.setText(String.valueOf(price));
+        int i = Integer.parseInt(mSettings.getString(APP_PREFERENCES_USERCOUNT, ""));
+        i = i + 1;
+        final SharedPreferences.Editor editor = mSettings.edit();
+        editor.putString(APP_PREFERENCES_USERCOUNT, String.valueOf(i));
+        editor.apply();
+
+        String username = (String) documentInfo.getFields().get(fields.getUsernameFields());
+
+
+        holder.UsernameItem.setText(username);
+        holder.UserCountItem.setText(number);
 
 
         view.setOnClickListener(new View.OnClickListener() {
@@ -143,15 +157,15 @@ public class RecyclerAdapterMenu extends BaseAdapter {
     }
 
     static class ViewHolder {
-        private TextView PeriodItem;
-        private TextView TypesOfTrainingItem;
-        private TextView PriceItem;
+        private TextView UsernameItem;
+        private TextView UserCountItem;
+
         //@BindView(R.id.label) TextView tvStoredItemName;;
 
         public ViewHolder(View view) {
-            PeriodItem = (TextView) view.findViewById(R.id.period);
-            TypesOfTrainingItem = (TextView) view.findViewById(R.id.types_of_training);
-            PriceItem = (TextView) view.findViewById(R.id.price);
+            UsernameItem = (TextView) view.findViewById(R.id.username);
+            UserCountItem = (TextView) view.findViewById(R.id.number);
+
         }
     }
 }
